backend/requirements.txt
fastapi
uvicorn
pydantic
httpx
aiofiles
pyyaml
pytest
rich         # optional CLI beautification
openai>=1.0.0
